# cc-http-service/openapi.json
{"openapi":"3.1.0","info":{"title":"Claude Code HTTP Service","description":"把容器内的 Claude Code 暴露为 HTTP/SSE 接口","version":"1.0.0"},"paths":{"/health":{"get":{"tags":["health"],"summary":"Health","operationId":"health_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"additionalProperties":true,"type":"object","title":"Response Health Health Get"}}}}}}},"/whoami":{"get":{"tags":["health"],"summary":"Whoami","description":"仅用于测试 token 是否生效。","operationId":"whoami_whoami_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"additionalProperties":true,"type":"object","title":"Response Whoami Whoami Get"}}}}},"security":[{"HTTPBearer":[]}]}},"/v1/query":{"post":{"tags":["query"],"summary":"Query Sync","description":"阻塞模式:等所有事件跑完,把最后一条 result 事件作为响应返回。\n\n适合一次性脚本调用,简单省事。","operationId":"query_sync_v1_query_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/QueryRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"HTTPBearer":[]}]}},"/v1/query/stream":{"post":{"tags":["query"],"summary":"Query Stream","description":"SSE 流式:边生成边推。","operationId":"query_stream_v1_query_stream_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/QueryRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"HTTPBearer":[]}]}},"/v1/files":{"post":{"tags":["query"],"summary":"Upload File","description":"上传文件(图片等)到容器内固定目录,返回容器内可访问的绝对路径。\n\n拿到 path 后,在 query 请求的 images 字段里传这个 path 即可让 cc 看到。","operationId":"upload_file_v1_files_post","requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_upload_file_v1_files_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/FileUploadResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"HTTPBearer":[]}]}},"/v1/sessions/create":{"post":{"tags":["sessions"],"summary":"Create Session","description":"创建一个常驻 session,返回 session_id。\n\nsession 在服务端维持一个 ClaudeSDKClient,后续 send 通过同一 session 续接上下文。","operationId":"create_session_v1_sessions_create_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SessionCreateRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"additionalProperties":true,"type":"object","title":"Response Create Session V1 Sessions Create Post"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"HTTPBearer":[]}]}},"/v1/sessions/list":{"get":{"tags":["sessions"],"summary":"List Sessions","operationId":"list_sessions_v1_sessions_list_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"additionalProperties":true,"type":"object","title":"Response List Sessions V1 Sessions List Get"}}}}},"security":[{"HTTPBearer":[]}]}},"/v1/sessions/{session_id}/send":{"post":{"tags":["sessions"],"summary":"Send Message","operationId":"send_message_v1_sessions__session_id__send_post","security":[{"HTTPBearer":[]}],"parameters":[{"name":"session_id","in":"path","required":true,"schema":{"type":"string","title":"Session Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SessionSendRequest"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"object","additionalProperties":true,"title":"Response Send Message V1 Sessions  Session Id  Send Post"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/v1/sessions/{session_id}/events":{"get":{"tags":["sessions"],"summary":"Stream Events","description":"订阅 session 的事件流。SSE 长连接,客户端边 send 边 listen。","operationId":"stream_events_v1_sessions__session_id__events_get","security":[{"HTTPBearer":[]}],"parameters":[{"name":"session_id","in":"path","required":true,"schema":{"type":"string","title":"Session Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/v1/sessions/{session_id}":{"delete":{"tags":["sessions"],"summary":"Close Session","operationId":"close_session_v1_sessions__session_id__delete","security":[{"HTTPBearer":[]}],"parameters":[{"name":"session_id","in":"path","required":true,"schema":{"type":"string","title":"Session Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"object","additionalProperties":true,"title":"Response Close Session V1 Sessions  Session Id  Delete"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"Body_upload_file_v1_files_post":{"properties":{"file":{"type":"string","contentMediaType":"application/octet-stream","title":"File"}},"type":"object","required":["file"],"title":"Body_upload_file_v1_files_post"},"FileUploadResponse":{"properties":{"filename":{"type":"string","title":"Filename"},"path":{"type":"string","title":"Path"},"size":{"type":"integer","title":"Size"}},"type":"object","required":["filename","path","size"],"title":"FileUploadResponse"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"HookMatcher":{"properties":{"matcher":{"type":"string","title":"Matcher","default":"*"},"command":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Command","description":"本地可执行文件路径,SDK 调用它处理 hook"},"timeout":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Timeout"}},"type":"object","title":"HookMatcher","description":"SDK hooks 配置。\n\n透传给 SDK,SDK 要求 hooks 是 Python 函数列表(非简单字符串 action)。\nHTTP API 暂时只接受 ``command`` 形式的回调:\n    {\n      \"matcher\": \"Write|Edit\",\n      \"command\": \"/usr/local/bin/my-hook.sh\"\n    }\nSDK 端的实现细节参见 _build_hooks。"},"HooksConfig":{"properties":{"PreToolUse":{"anyOf":[{"items":{"$ref":"#/components/schemas/HookMatcher"},"type":"array"},{"type":"null"}],"title":"Pretooluse"},"PostToolUse":{"anyOf":[{"items":{"$ref":"#/components/schemas/HookMatcher"},"type":"array"},{"type":"null"}],"title":"Posttooluse"},"Stop":{"anyOf":[{"items":{"$ref":"#/components/schemas/HookMatcher"},"type":"array"},{"type":"null"}],"title":"Stop"},"SubagentStop":{"anyOf":[{"items":{"$ref":"#/components/schemas/HookMatcher"},"type":"array"},{"type":"null"}],"title":"Subagentstop"}},"type":"object","title":"HooksConfig"},"McpServerConfig":{"properties":{"type":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Type"},"command":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Command"},"args":{"anyOf":[{"items":{"type":"string"},"type":"array"},{"type":"null"}],"title":"Args"},"env":{"anyOf":[{"additionalProperties":{"type":"string"},"type":"object"},{"type":"null"}],"title":"Env"},"url":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Url"},"headers":{"anyOf":[{"additionalProperties":{"type":"string"},"type":"object"},{"type":"null"}],"title":"Headers"}},"type":"object","title":"McpServerConfig","description":"MCP 服务器配置,支持 stdio/http/sse 三种传输。"},"QueryRequest":{"properties":{"prompt":{"type":"string","title":"Prompt","description":"用户消息"},"work_dir":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Work Dir","description":"Claude Code 工作目录,默认 /home/vscode/cc-home"},"model":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Model"},"permission_mode":{"anyOf":[{"type":"string","enum":["default","acceptEdits","plan","bypassPermissions","dontAsk","auto"]},{"type":"null"}],"title":"Permission Mode"},"allowed_tools":{"anyOf":[{"items":{"type":"string"},"type":"array"},{"type":"null"}],"title":"Allowed Tools"},"system_prompt":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"System Prompt"},"hooks":{"anyOf":[{"$ref":"#/components/schemas/HooksConfig"},{"type":"null"}],"description":"生命周期 hook。command 形式:SDK 会 spawn 进程传 stdin JSON"},"agents":{"anyOf":[{"additionalProperties":{"$ref":"#/components/schemas/SubagentDef"},"type":"object"},{"type":"null"}],"title":"Agents"},"mcp_servers":{"anyOf":[{"additionalProperties":{"$ref":"#/components/schemas/McpServerConfig"},"type":"object"},{"type":"null"}],"title":"Mcp Servers"},"images":{"anyOf":[{"items":{"type":"string"},"type":"array"},{"type":"null"}],"title":"Images","description":"容器内图片绝对路径列表(图片需先放到挂载目录里,或通过 /v1/files 上传)"},"extra_options":{"anyOf":[{"additionalProperties":true,"type":"object"},{"type":"null"}],"title":"Extra Options"}},"type":"object","required":["prompt"],"title":"QueryRequest"},"SessionCreateRequest":{"properties":{"work_dir":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Work Dir"},"model":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Model"},"permission_mode":{"anyOf":[{"type":"string","enum":["default","acceptEdits","plan","bypassPermissions","dontAsk","auto"]},{"type":"null"}],"title":"Permission Mode"},"allowed_tools":{"anyOf":[{"items":{"type":"string"},"type":"array"},{"type":"null"}],"title":"Allowed Tools"},"system_prompt":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"System Prompt"}},"type":"object","title":"SessionCreateRequest"},"SessionSendRequest":{"properties":{"message":{"type":"string","title":"Message"},"images":{"anyOf":[{"items":{"type":"string"},"type":"array"},{"type":"null"}],"title":"Images"}},"type":"object","required":["message"],"title":"SessionSendRequest"},"SubagentDef":{"properties":{"description":{"type":"string","title":"Description"},"prompt":{"type":"string","title":"Prompt"},"tools":{"anyOf":[{"items":{"type":"string"},"type":"array"},{"type":"null"}],"title":"Tools"},"model":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Model"}},"type":"object","required":["description","prompt"],"title":"SubagentDef"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}},"securitySchemes":{"HTTPBearer":{"type":"http","scheme":"bearer"}}}}